+++ visualization/app/codeCharta/state/store/appSettings/labelsPerMap/labelsPerMap.actions.ts
@@ -1,0 +1,3 @@
+import { createAction, props } from "@ngrx/store"
+
+export const setLabelsPerMap = createAction("SET_LABELS_PER_MAP", props<{ value: boolean }>())